+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cpp: Limit rate of SynchronousClock skip warnings and report frequency drop
 - envs: Allow gain scales larger than one in servo environments
 - envs: Drop Gymnasium environment versioning
+- examples: Allow users to select the Gymnasium environment in MPC example
 - spines: Factor common observers and sensors between spines
 - utils: Generalize `abs_bounded_derivative_filter` to vector inputs
 - utils: Generalize `bounded_derivative_filter` to vector inputs